+++ superformula-frontend/src/domains/users/components/EditModal/EditModal.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/mixins.scss';
+@import 'styles/mixins.scss';
 
 .background {
     display: flex;
