+++ superformula-frontend/src/domains/users/components/EditModal/EditModal.stories.tsx
@@ -2,12 +2,12 @@ import React from 'react'
 import { Story } from '@storybook/react'
 import { Meta } from '@storybook/react/types-6-0'
 
-import styles from '../../../styles/themes.module.scss'
-import UserCard, { UserCardProps } from './UserCard'
+import styles from 'styles/themes.module.scss'
+import EditModal, { EditModalProps } from './EditModal'
 
 export default {
-    title: 'Components/UserManagement/Card',
-    component: UserCard,
+    title: 'Components/UserManagement/EditModal',
+    component: EditModal,
     decorators: [
         (StoryComponent) => (
             <div className={styles.defaultTheme}>
@@ -17,7 +17,7 @@ export default {
     ],
 } as Meta
 
-const Template: Story<UserCardProps> = (args: UserCardProps) => <UserCard {...args} />
+const Template: Story<EditModalProps> = (args: EditModalProps) => <EditModal {...args} />
 
 export const Main = Template.bind({})
 Main.args = {
