+++ superformula-frontend/src/domains/users/components/UserCard/UserCard.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/mixins.scss';
+@import 'styles/mixins.scss';
 
 .card {
     text-align: center;
