+++ superformula-frontend/src/domains/users/components/UserCard/index.tsx
