+++ superformula-frontend/src/domains/users/models/User.tsx
