+++ superformula-frontend/src/domains/users/pages/UserManagement/UserManagement.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles//mixins.scss';
+@import 'styles//mixins.scss';
 
 .page {
     overflow: hidden;
