+++ superformula-frontend/src/domains/users/pages/UserManagement/UserManagement.tsx
@@ -1,12 +1,12 @@
 import React, { ReactElement, useState } from 'react'
-import UserCard from '../../userManagement/components/UserCard/UserCard'
-import Input from '../../components/Input/Input'
+import UserCard from 'domains/users/components/UserCard/UserCard'
+import Input from 'components/Input/Input'
+import Button from 'components/Button/Button'
+import useQueryParams, { QueryParam } from 'hooks/useQueryParams'
+import Loader from 'components/Loader/Loader'
+import useListUsers from 'hooks/useListUsers'
+import { getIntegerQueryParam } from 'utils/queryHelper'
 import styles from './UserManagement.module.scss'
-import Button from '../../components/Button/Button'
-import useQueryParams, { QueryParam } from '../../hooks/useQueryParams'
-import Loader from '../../components/Loader/Loader'
-import useListUsers from '../../hooks/useListUsers'
-import { getIntegerQueryParam } from '../../utils/queryHelper'
 
 const image =
     'https://media.istockphoto.com/photos/funny-west-highland-white-terrier-dog-decorated-with-photo-props-sits-picture-id1292884801'
